+++ examples/src/water-effect.js
@@ -19,13 +19,14 @@ app.loader.add("water_dudv.jpg", "assets/textures/water_dudv.jpg")
 let sprite1 = new PIXI3D.PostProcessingSprite(app.renderer)
 let sprite2 = new PIXI3D.PostProcessingSprite(app.renderer)
 
-let pipeline = PIXI3D.StandardPipeline.from(app.renderer)
+let pipeline = app.renderer.plugins.pipeline
 
 // Set the render texture for the default material render pass.
 pipeline.materialPass.renderTexture = sprite1.renderTexture
 
 // Create a new water render pass and set the render texture
-let waterPass = pipeline.addRenderPass(new PIXI3D.MaterialRenderPass(app.renderer, "water"))
+let waterPass = new PIXI3D.MaterialRenderPass(app.renderer, "water")
+pipeline.renderPasses.push(waterPass)
 waterPass.renderTexture = sprite2.renderTexture
 
 app.loader.load((loader, resources) => {